Create new file structure and rename extensions to .jsx
