+++ app/components/SideBar/Bar/MenuBottom/LastBlockTime/LastBlockTime.jsx
@@ -4,12 +4,19 @@ import { FormattedRelative } from "shared";
 import { useLastBlockTime } from "./hooks";
 
 function LastBlockTime({ lastBlockTimestamp, clearTimeout, setTimeout }) {
-  const { state } = useLastBlockTime(lastBlockTimestamp, clearTimeout, setTimeout);
+  const {
+    lastBlockDate,
+    lastBlockIsRecent
+  } = useLastBlockTime(
+    lastBlockTimestamp,
+    clearTimeout,
+    setTimeout
+  );
 
-  return state && state.lastBlockDate ?
-    state.lastBlockIsRecent ?
+  return lastBlockDate ?
+    lastBlockIsRecent ?
       <T id="sidebar.lastBlockIsRecent" m="< 1 minute ago" /> :
-      <FormattedRelative value={state.lastBlockDate} updateInterval={1 * 1000} />
+      <FormattedRelative value={lastBlockDate} updateInterval={1 * 1000} />
     : null;
 
 }
