+++ app/components/SideBar/SideBar.jsx
@@ -1,8 +1,12 @@
-import Bar from "./Bar";
+import Bar from "./Bar/Bar";
 import { useSideBar } from "./hooks";
 
 function SideBar() {
-  const { isShowingAccounts, onShowAccounts, onHideAccounts } = useSideBar();
+  const {
+    isShowingAccounts,
+    onShowAccounts,
+    onHideAccounts
+  } = useSideBar();
 
   return (
     <Bar
