+++ app/components/SideBar/SideBar.module.css
@@ -1,15 +1,3 @@
-.hamburger {
-	height: 20px;
-	width: 20px;
-	background-repeat: no-repeat;
-	background-image: var(--menu-hamburger);
-	cursor: pointer;
-}
-
-.hamburger:hover {
-	opacity: 0.7;
-}
-
 .sidebarReduceArrow {
 	height: 20px;
 	width: 20px;
@@ -42,19 +30,6 @@
 	flex-direction: row;
 }
 
-.sidebarLogo,
-.reducedSidebarLogo {
-	display: flex;
-	position: relative;
-	z-index: 3;
-	padding: 20px 18px 20px 18px;
-	border-bottom: 1px solid var(--white-border);
-}
-
-.reducedSidebarLogo {
-	flex-direction: column;
-}
-
 .sidebarMain {
 	position: relative;
 	flex: 1;
@@ -68,56 +43,6 @@
 	overflow-y: auto;
 	overflow-x: hidden;
 	flex: 1;
-}
-
-.sidebarMenuTotalBalanceExtended {
-	z-index: 2;
-	position: absolute;
-	top: 0;
-	height: 100%;
-	width: 100%;
-	flex-direction: column;
-	justify-content: flex-end;
-	background-color: var(--background-container-opaque);
-	transition: all 100ms cubic-bezier(0.86, 0, 0.07, 1);
-}
-
-.sidebarMenuTotalBalanceExtendedBottom {
-	padding-right: 18px;
-}
-
-.sidebarMenuTotalBalanceExtendedBottomAccount {
-	height: 44px;
-	display: flex;
-	padding-left: 52px;
-	justify-content: space-between;
-	background-image: var(--accounts-default);
-	background-size: 20px auto;
-	background-repeat: no-repeat;
-	background-position: 19px 0;
-}
-
-.sidebarMenuTotalBalanceExtendedBottomAccount.imported {
-	background-image: var(--accounts-imported);
-}
-
-.sidebarMenuTotalBalanceExtendedBottomAccountName {
-	color: var(--stroke-color-hovered);
-	font-size: 13px;
-	overflow: hidden;
-	white-space: nowrap;
-	text-overflow: ellipsis;
-}
-
-.sidebarMenuTotalBalanceExtendedBottomAccountNumber {
-	font-family: var(--font-family-regular);
-	color: var(--input-color-default);
-	font-size: 11pt;
-	font-weight: 600;
-}
-
-.sidebarMenuTotalBalanceExtendedBottomAccountNumber .balanceSmall {
-	font-size: 11pt;
 }
 
 .sidebarMenuBottomTotalBalanceShort {
@@ -230,192 +155,6 @@
 	margin-bottom: 10px;
 }
 
-.sidebarMenuLinkNotificationIcon {
-	background-repeat: no-repeat;
-	background-image: var(--bell-notification-icon);
-	background-position: center;
-	background-size: 10px;
-	width: 14px;
-	height: 14px;
-	margin-left: 34px;
-	margin-top: 10px;
-	position: absolute;
-	display: inline-block;
-	z-index: 10;
-	border-radius: 50%;
-	background-color: var(--stroke-color-focused);
-}
-
-@keyframes spin {
-	from {
-		transform: rotate(0deg);
-	}
-
-	to {
-		transform: rotate(360deg);
-	}
-}
-
-.rescan-button.spin {
-	animation: spin 1s linear infinite;
-}
-
-.menuLink {
-	display: flex;
-	position: relative;
-	z-index: 1;
-	align-items: center;
-	cursor: pointer;
-	height: 52px;
-	padding-left: 65px;
-	color: var(--sidebar-menu-link);
-	font-size: 15px;
-	text-decoration: none;
-	transition: color 0.13s ease-in-out;
-	background-position: 20px 50%;
-	background-size: 20px;
-	background-repeat: no-repeat;
-	user-select: none;
-}
-
-.menuLink.settingsIcon {
-	background-image: var(--menu-settings-default);
-}
-
-.menuLink.settingsIcon:not(.menuLinkActive):hover {
-	background-image: var(--menu-animation-settings);
-}
-
-.menuLink.helpIcon {
-	background-image: var(--menu-help-default);
-}
-
-.menuLink.helpIcon:not(.menuLinkActive):hover {
-	background-image: var(--menu-animation-help);
-}
-
-.menuLink.ticketsIcon {
-	background-image: var(--menu-tickets-default);
-}
-
-.menuLink.ticketsIcon:not(.menuLinkActive):hover {
-	background-image: var(--menu-animation-tickets);
-}
-
-.menuLink.transactionsIcon {
-	background-image: var(--menu-transactions-default);
-}
-
-.menuLink.trezorIcon {
-	background-image: var(--menu-trezor-default);
-}
-
-.menuLink.trezorIcon:not(.menuLinkActive):hover {
-	background-image: var(--menu-trezor-hover);
-}
-
-.menuLink.transactionsIcon:not(.menuLinkActive):hover {
-	background-image: var(--menu-animation-transactions);
-}
-
-.menuLink.accountsIcon {
-	background-image: var(--menu-accounts-default);
-}
-
-.menuLink.accountsIcon:not(.menuLinkActive):hover {
-	background-image: var(--menu-animation-accounts);
-	background-size: 22px;
-	background-position: 19px 50%;
-}
-
-.menuLink.securitycntrIcon {
-	background-image: var(--menu-securitycntr-default);
-}
-
-.menuLink.securitycntrIcon:not(.menuLinkActive):hover {
-	background-image: var(--menu-animation-security);
-}
-
-.menuLink.overviewIcon {
-	background-image: var(--menu-overview-default);
-}
-
-.menuLink.overviewIcon:not(.menuLinkActive):hover {
-	background-image: var(--menu-animation-overview);
-}
-
-.menuLink.governanceIcon {
-	background-image: var(--menu-governance-default);
-}
-
-.menuLink.governanceIcon:not(.menuLinkActive):hover {
-	background-image: var(--menu-animation-governance);
-}
-
-.menuLink.lnIcon {
-	background-image: var(--menu-ln-default);
-}
-
-.menuLink.lnIcon:not(.menuLinkActive):hover {
-	background-image: var(--menu-ln-hover);
-}
-
-.menuLink:hover {
-	color: var(--input-color-default);
-}
-
-.menuLinkActive {
-	color: var(--input-color-default);
-}
-
-.menuLinkActive.settingsIcon {
-	background-image: var(--menu-settings-active);
-	cursor: default;
-}
-
-.menuLinkActive.helpIcon {
-	background-image: var(--menu-help-active);
-	cursor: default;
-}
-
-.menuLinkActive.ticketsIcon {
-	background-image: var(--menu-tickets-active);
-	cursor: default;
-}
-
-.menuLinkActive.transactionsIcon {
-	background-image: var(--menu-transactions-active);
-	cursor: default;
-}
-
-.menuLinkActive.accountsIcon {
-	background-image: var(--menu-accounts-active);
-	cursor: default;
-}
-
-.menuLinkActive.securitycntrIcon {
-	background-image: var(--menu-securitycntr-active);
-	cursor: default;
-}
-
-.menuLinkActive.overviewIcon {
-	background-image: var(--menu-overview-active);
-	cursor: default;
-}
-
-.menuLinkActive.governanceIcon {
-	background-image: var(--menu-governance-active);
-	cursor: default;
-}
-
-.menuLinkActive.trezorIcon {
-	background-image: var(--menu-trezor-active);
-}
-
-.menuLinkActive.lnIcon {
-	background-image: var(--menu-ln-active);
-}
-
 .menuCaret {
 	position: absolute;
 	height: 52px;
@@ -439,13 +178,6 @@
 		flex-direction: row-reverse;
 	}
 
-	.menuLink {
-		height: 70px;
-		width: 76px;
-		padding-left: 0;
-		background-position: 50%;
-	}
-
 	.menuCaret {
 		width: 76px;
 		height: 65px;
@@ -465,4 +197,18 @@
 	.sidebarMenuBottomLatestBlock .tooltipContainer span {
 		visibility: hidden;
 	}
+}
+
+@keyframes spin {
+	from {
+		transform: rotate(0deg);
+	}
+
+	to {
+		transform: rotate(360deg);
+	}
+}
+
+.rescan-button.spin {
+	animation: spin 1s linear infinite;
 }
