+++ app/containers/Wallet.js
@@ -17,7 +17,7 @@ import GovernancePage from "components/views/GovernancePage/GovernancePage";
 import ProposalDetailsPage from "components/views/ProposalDetailsPage/ProposalDetailsPage";
 import TrezorPage from "components/views/TrezorPage";
 import LNPage from "components/views/LNPage";
-import SideBar from "components/SideBar";
+import SideBar from "components/SideBar/SideBar";
 import { BlurableContainer } from "layout";
 import { walletContainer, theming } from "connectors";
 import ReactTimeout from "react-timeout";
